Fix an issue where the extension could get stuck in a bad state.
Not using the extension for several days seems to get the session token into a weird
state where it isn't timed out, but it also (erroneously) doesn't have permissions.
Restarting the browser or the extension can fix it, but this commit bandaids the
issue by catching the issue when polling for tasks and invalidating the session to
force another login. This isn't foolproof (this error can be hit while performing
other actions too) but should catch the vast majority of error states.