Fix a regression where HEAD requests would be sent to the wrong URL. Fixes #125.
In c09fa9e340b6a3a3d9a15b42f51a8c8fdd1b757e, getMetadataFileType was factored out
but the logic was erroneously changed to use the URL without the query string to
make the request. The URL without query string was intended only to to aid in
guessing the file type or file name, not to change where the request actually
goes.